+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
  - (Added) A ``setdefault`` function that can be used in the preamble and actions of a statechart to assign default values to variables.
  - (Changed) Meta-Event *step started* has a ``time`` attribute.
  - (Changed) The current event queue can be consulted as a list using ``interpreter._event_queue``. 
+ - (Fixed) Internal events are processed before external ones (regression introduced in 1.3.0).
  - (Fixed) Hook-errors reported by ``sismic-bdd`` CLI are a little bit more verbose (`#81 <https://github.com/AlexandreDecan/sismic/issues/81>`__).
  - (Fixed) Optional transition for ``testing.transition_is_processed``, as previously promised by its documentation but not implemented.
  - (Removed) Internal module ``sismic.interpreter.queue``.